Set github actions fail fast to false
There is an ongoing issue with Rustup 1.24 which doesn't handle systems with a limited memory pool very well (like the CI). Since cargo-msrv uses rustup to install toolchains, the CI fails all the time. With this commit I aim to find out whether the issue only holds true for the msvc platform or also for the other platforms. We may need to downgrade and lock our rustup version for the CI.