Add tracing to enable logging debugging information without anoying users
Adds logic to initialize tracing. The tracing information is logged to a file, which will be rotated daily.
The events logged using tracing may be more detailed, or simply different from the output we would like to show users.

Tracing may be disabled by running cargo-msrv with the `--no-log` flag. In that case, not even a log file will be created, which may be useful if the log file location (which is currently not configurable) is inaccessible.

For the implementation it is important that we always collect the guards of the non-blocking tracing appender(s), since they will only live as long as the lifetime of the worker guard. If we don't do this, the guards would be dropped after the scope of `if !config.no_tracing() { ... }` ended, and as a result, anything logged in `init_and_run` would not be logged.