Ignore chrono and time both potentially segfault audit reports
Time crate has been updated, but we depend on chrono and are waiting for it to update. The PR is open but the maintainer seems unresponsive. For now, we'll wait a bit as it's actually very unlikely for the segfault to take place, and, the impact on cargo-msrv is not significant.
In time, we may want to replace chrono, assuming it's unmaintained.