Parse bare version numbers left-to-right
Previously we would split the string and do some magic on each of the parts. Now we'll instead try to properly parse the version number. Currently, this only includes the bare version we already supported, but this change makes it much easier to parse the pre-release identifier as defined by the semver spec.