Install the latest Rust version for release build action
This is necessary since cargo about already updated to rust 2021, so to compile it we need at least Rust version 1.56, while so far Rust version 1.55 was available.

In addition, taking the latest stable may add new optimizations not available on earlier versions.

A disadvantage is the extra added toolchain install time.

Also removed the rust-cache since it never really had any hits, and to ensure it isn't interfering with the toolchain version.