Fix verify subcommand would not allow a custom check command to be set
Prior to adding verify as a subcommand, and the deprecation of the --verify flag, the custom check command would be matched by the arg matcher of the root command. Now that verify has been added as its own subcommand, the custom check command does not seem to be inherited from the root command. This was fixed by defining the custom check command on the verify subcommand in addition to the root command.