Change default check command to `cargo check`
Previously, the default check command was `cargo check --all`.
This worked if you wanted to find the MSRV of the greatest common MSRV of all packages in the workspace. However, if you want to find the MSRV of the current package (whether it's a virtual workspace, workspace root, or regular package), finding the greatest common MSRV is not the goal of the search for the MSRV. Instead, you'll want to define the MSRV for the smallest logical unit. In this case, when a package depends on workspace packages (or external packages), it's MSRV will already be dependent on these packages with the regular `cargo check` command. As such it makes sense to change the default to be just `cargo check`.

To revert to the old behaviour, you can run cargo-msrv with `cargo msrv -- cargo check --all`.

BREAKING CHANGE: Changed default cargo-msrv (find) check command from `cargo check --all` to `cargo check`

Co-authored-by: Christopher Schramm  <cschramm@users.noreply.github.com>