Port cargo-msrv to attribute macro variant of Clap v3
The port consists of two important parts:

1) Add types and parsing for all inputs, and allow Clap to use these parsers to convert directly to the types we define as input
2) Translate from the cli to  Config

The first part is the biggest win, as it will give a higher guarantee on well-formed input.

The second part, also is a small win, as it used to be one god function, with lots of conditions around builders. The way the Config is set up is still the same (to prevent refactoring multiple factors at once), but the translation functions which check conditions and set builder values are now better separated into their own functions.

The global config struct is problematic because it doesn't scale too well with the many options and subcommands.

The typed structure on which the Clap attribute macro is built, would actually be a decent config replacement, however using it directly would mean that the CargoCli struct is special, as it would be both a frontend (via generated Clap attributes and its structure) and the config itself, with an absolute requirement to stay in sync.

At one point in the future, we should work towards layering the config, which would allow environment variable overrides, and config files. For this commit however, it would be too much change, and this commit by itself is already very large.