add feature toggle for 'rust-release-dist-source', default on
This is a non-breaking change which allows users to install with

```
cargo install --no-default-features cargo-msrv
```

and more quickly with less dependencies. For me it took 1m02s
for the full build and 41s with `--no-default-features`.

Since the 'changelog' release channel is the default,
there is no differnece in typical usage.

The binary size shrinks from 9.5MB to 4.4MB as well.