Expand accepted version ranges for --min and --max
Previously, --min and --max would strictly take semver versions. With this commit, we are more liberal, and also allow two component semver versions.

Notably, the max version may be too liberal, as given version 1.56.1 will now be accepted if the max version is 1.56.0. This is accepted as a 'workaround', which allows the version 1.56.1 to match when a given version spec is `1.56`.