Be more precise in version ranges accepted by --min and --max
In the previous commit, the version taken by --min and --max would already be a bare version, but the config still used a strict semver version. As a result, a workaround had to be applied to accept two component bare versions where the version compared to, had a patch version > 0. As an example, this allowed the version 1.56.1 to match when a given version spec is `1.56`.
However, this also resulted in a (known) regression, where a full semver version 1.56.0 still matched a given version spec 1.56.1 (patch version 1 > patch version 0, which was an unintended, but known side effect of the previous version of the implementation).

Now we replaced the min and max versions in the config to be bare versions, and implemented two comparators (as regular methods, as the Eq, Ord and PartialOrd contract would not be upheld) accordingly: BareVersion::is_at_least and BareVersion::is_at_most.