Add option which allows users to specify the rust-version to be verified
Previously the rust-version (MSRV) could only be verified if it was set in the Cargo manifest. This required projects using older versions to use the non-standard package.metadata.msrv field. In addition, it required projects where we just want to run verify once or twice, to set the version temporarily in the Cargo manifest.
Now this is no longer required, as one can instead simply run cargo verify --rust-version <ARG>. The argument will be prioritized over the version specified in the Cargo manifest.

NB: The deprecated cargo `msrv --verify` does not support this option. The dedicated `cargo msrv verify` subcommand must be used.