Fix case where unrooted relative manifest-path would fail
The program would fail because an unrooted relative manifest-path does not have a parent. We rely on the Path::parent() method to find the crate directory.
The fix first canonicalizes the path. This assumes no cargo crate is rooted at the top of a file system. Possibly, if you mount a crate to a separate file system, it will still fail, but it is assumed that this is not the common case. If we need to support such a case, we no longer can use the Path::parent() method. We may check if manifest_path == "Cargo.toml", and then accept "" or "." as crate root.