Refactor final output of a subcommand to use a specific event
The event is called SubcommandResult, and is composed of the specific output of the subcommand.

This event for the output of a subcommand makes it more clear where to look for the final result.
When using the json output format, you can now simply look for the final result, instead of matching more specific events.