Rename events for consistency and to be more clearly identifiable
Renamed "Action" to "SubcommandInit" to better represent that the reported subcommand id matches the now running subcommand.

Renamed "CompatibilityCheckMethod" to "CheckMethod" to be more consistent within the family of "check" events. Internally the event was already renamed.

Renamed "Compatibility" to "CheckResult" to be more consistent within the family of "check" events. Internally the event was already renamed.