Refactor Compatibility report to just report the error if there is any
Previously, we would also just report the error; but wrapped in a "report" object. For instances where we wanted to report the success, the report would be an empty object. The name of the object would be conditional on the success of the instance. In this commit we cleaned these things up, by removing the inner, conditionally named object, and instead adding an optional error field to the Compatibility type.