feat: Add method to check we're dealing with a Cargo project
Currently, cargo msrv requires Rust projects to make use of Cargo. But if the given path was not a Cargo project, we would return a generic IO error that 'Cargo.toml' does not exist. This feature is the basis necessary, to streamline the user output, and inform users that the given project is not a Cargo project.