infra: Remove nightly from ci-test workflow
Output of cargo msrv tests is currently different on nightly, compared to stable (and beta). As such, some of our end-to-end tests were failing. Nightly is only a nice to have for now, and handling the output separately is not in my best interest regarding time management. As such, I decided to remove it.