Don't pass `OutputFormat` to `Context` anymore
It seems the only place where it was actually used was in `src/sub_command/verify.rs`.
And even there it was probably not needed since the reporter for 'no output'
would most likely discard the message anyway.

Not passing the OutputFormat to Context simplifies the code, and is probably
also cleaner, since the Context and the commands likely should not care about
the output format. The main code will set up reporting according to the output
format.