Add e2e test via Github workflow
In #1036, an issue was reported which reported toolchain install issues, possibly caused by path separator issues on windows (ubuntu and mac seem to be fine).

With this commit, we add a run of cargo-msrv as an e2e test to our own CI on windows and mac while we're at it.
Previously, we would only run on ubuntu; the e2e test was a side-effect of checking our own MSRV.
We do have tests which run via the 'cargo run' vehicle and temporary directories, but that's not as sandboxed as this approach.

Hopefully, running all primary platforms will (1) give an indication of whether #1036 is indeed a cargo-msrv issue, and (2) prevent similar bugs in the future.