Switch the `rust-releases-dist-source` off by default.
Justification: enabling this almost doubles the dependency count, takes
it from ~200 to ~400.
The `cargo test` results on my machine improved after this change.
The `rust-releases` crate also recommends using the
changelog source. The increase in compilation time is more than double,
because all the aws crates take a very long time to compile.